requirements.txt
asgiref==3.5.2
Django==4.0.6
djangorestframework==3.13.1
mysqlclient==2.1.1
pytz==2022.1
sqlparse==0.4.2
